Ensuring to require relevant context type
