+++ trueseeing/app/cmd/android/search.py
@@ -28,45 +28,7 @@ class SearchCommand(CommandMixin):
       '/dcx':dict(e=self._search_derived_class, n='/dcx class [pat]', d='search classes extending ones matching pattern'),
       '/dci':dict(e=self._search_implementing_class, n='/dci interface [pat]', d='search classes implementing interfaces matching pattern'),
       '/dm':dict(e=self._search_defined_method, n='/dm class [pat]', d='search classes defining methods matching pattern'),
-      '/f':dict(e=self._search_file, n='/f [pat]', d='search files those names matching pattern'),
-      '/s':dict(e=self._search_string, n='/s pat', d='search files for string'),
     }
-
-  async def _search_file(self, args: deque[str]) -> None:
-    self._helper.require_target()
-
-    _ = args.popleft()
-
-    if args:
-      pat = args.popleft()
-    else:
-      pat = '.'
-
-    context = await self._helper.get_context_analyzed('apk', level=1)
-    level = context.get_analysis_level()
-    if level < 3:
-      ui.warn('detected analysis level: {} ({}) -- try analyzing fully (\'aa\') to maximize coverage'.format(level, self._helper.decode_analysis_level(level)))
-    for path in context.store().query().file_find(pat=pat, regex=True):
-      ui.info(f'{path}')
-
-  async def _search_string(self, args: deque[str]) -> None:
-    self._helper.require_target()
-
-    _ = args.popleft()
-
-    if not args:
-      ui.fatal('need pattern')
-
-    pat = args.popleft()
-
-    ui.info('searching in files: {pat}'.format(pat=pat))
-
-    context = await self._helper.get_context_analyzed('apk', level=1)
-    level = context.get_analysis_level()
-    if level < 3:
-      ui.warn('detected analysis level: {} ({}) -- try analyzing fully (\'aa\') to maximize coverage'.format(level, self._helper.decode_analysis_level(level)))
-    for path in context.store().query().file_search(pat=pat.encode('latin1'), regex=True):
-      ui.info(f'{path}')
 
   async def _search_call(self, args: deque[str]) -> None:
     self._helper.require_target()
