+++ trueseeing/app/cmd/android/show.py
@@ -21,48 +21,11 @@ class ShowCommand(CommandMixin):
 
   def get_commands(self) -> CommandMap:
     return {
-      'pf':dict(e=self._show_file, n='pf[x][!] path [output.bin]', d='show file (x: hex)'),
-      'pf!':dict(e=self._show_file),
-      'pfx':dict(e=self._show_file),
-      'pfx!':dict(e=self._show_file),
       'pd':dict(e=self._show_disasm, n='pd[!] class [output.smali]', d='show disassembled class'),
       'pd!':dict(e=self._show_disasm),
       'pk':dict(e=self._show_solved_constant, n='pk op index', d='guess and show what constant would flow into the index-th arg of op (!: try harder)'),
       'pt':dict(e=self._show_solved_typeset, n='pt op index', d='guess and show what type would flow into the index-th arg of op'),
     }
-
-  async def _show_file(self, args: deque[str]) -> None:
-    outfn: Optional[str] = None
-
-    self._helper.require_target()
-
-    cmd = args.popleft()
-
-    if not args:
-      ui.fatal('need path')
-
-    path = args.popleft()
-
-    if args:
-      import os
-      outfn = args.popleft()
-      if os.path.exists(outfn) and not cmd.endswith('!'):
-        ui.fatal('outfile exists; force (!) to overwrite')
-
-    from binascii import hexlify
-
-    context = await self._helper.get_context_analyzed('apk', level=1)
-    level = context.get_analysis_level()
-    if level < 3:
-      ui.warn('detected analysis level: {} ({}) -- try analyzing fully (\'aa\') to maximize coverage'.format(level, self._helper.decode_analysis_level(level)))
-    d = context.store().query().file_get(path)
-    if d is None:
-      ui.fatal('file not found')
-    if outfn is None:
-      sys.stdout.buffer.write(d if 'x' not in cmd else hexlify(d))
-    else:
-      with open(outfn, 'wb') as f:
-        f.write(d if 'x' not in cmd else hexlify(d))
 
   async def _show_disasm(self, args: deque[str]) -> None:
     outfn: Optional[str] = None
