+++ trueseeing/app/cmd/search.py
@@ -1,0 +1,60 @@
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
+from collections import deque
+
+from trueseeing.core.model.cmd import CommandMixin
+from trueseeing.core.ui import ui
+
+if TYPE_CHECKING:
+  from trueseeing.api import CommandHelper, Command, CommandMap
+
+class SearchCommand(CommandMixin):
+  def __init__(self, helper: CommandHelper) -> None:
+    self._helper = helper
+
+  @staticmethod
+  def create(helper: CommandHelper) -> Command:
+    return SearchCommand(helper)
+
+  def get_commands(self) -> CommandMap:
+    return {
+      '/f':dict(e=self._search_file, n='/f [pat]', d='search files those names matching pattern'),
+      '/s':dict(e=self._search_string, n='/s pat', d='search files for string'),
+    }
+
+  async def _search_file(self, args: deque[str]) -> None:
+    self._helper.require_target()
+
+    _ = args.popleft()
+
+    if args:
+      pat = args.popleft()
+    else:
+      pat = '.'
+
+    context = await self._helper.get_context_analyzed('file', level=1)
+    level = context.get_analysis_level()
+    if level < 3:
+      ui.warn('detected analysis level: {} ({}) -- try analyzing fully (\'aa\') to maximize coverage'.format(level, self._helper.decode_analysis_level(level)))
+    for path in context.store().query().file_find(pat=pat, regex=True):
+      ui.info(f'{path}')
+
+  async def _search_string(self, args: deque[str]) -> None:
+    self._helper.require_target()
+
+    _ = args.popleft()
+
+    if not args:
+      ui.fatal('need pattern')
+
+    pat = args.popleft()
+
+    ui.info('searching in files: {pat}'.format(pat=pat))
+
+    context = await self._helper.get_context_analyzed('file', level=1)
+    level = context.get_analysis_level()
+    if level < 3:
+      ui.warn('detected analysis level: {} ({}) -- try analyzing fully (\'aa\') to maximize coverage'.format(level, self._helper.decode_analysis_level(level)))
+    for path in context.store().query().file_search(pat=pat.encode('latin1'), regex=True):
+      ui.info(f'{path}')
